Moving test to mirror implementation path
