+++ src/test/java/com/github/junrar/JunrarTests.java
@@ -1,7 +1,8 @@
-package com.github.junrar.functional;
+package com.github.junrar;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 
 import java.io.File;
 import java.io.IOException;
@@ -11,12 +12,9 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import com.github.junrar.Junrar;
-import com.github.junrar.TestCommons;
 import com.github.junrar.exception.RarException;
 
-public class ExtractionTests {
-	
+public class JunrarTests {
 	private static File tempFolder;
 	
 	@BeforeClass
@@ -38,5 +36,28 @@ public class ExtractionTests {
 		File fooDir = new File(tempFolder,"foo");
 		assertTrue(fooDir.exists());		
 		assertEquals("baz\n", FileUtils.readFileToString(new File(fooDir,"bar.txt")));
+	}
+	
+	@Test
+	public void ifIsDirInsteadOfFile_ThrowException() throws RarException, IOException {
+		try {
+			Junrar.extract(tempFolder, tempFolder);
+		}catch (IllegalArgumentException e) {
+			assertEquals("First argument should be a file but was "+tempFolder.getAbsolutePath(), e.getMessage());
+			return;
+		}
+		fail();		
+	}
+	
+	@Test
+	public void ifIsFileInsteadOfDir_ThrowException() throws RarException, IOException {
+		File rarFileOnTemp = TestCommons.writeTestRarToFolder(tempFolder);
+		try {
+			Junrar.extract(rarFileOnTemp, rarFileOnTemp);
+		}catch (IllegalArgumentException e) {
+			assertEquals("Second argument should be a directory but was "+rarFileOnTemp.getAbsolutePath(), e.getMessage());
+			return;
+		}
+		fail();		
 	}
 }
